# backend/api-docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CloudPhone Widget Backend API",
    "version": "v1.3.0",
    "description": "API for the backend of the custom CloudPhone widget."
  },
  "servers": [
    {
      "url": "/",
      "description": "目前開啟 Swagger 的 API 伺服器（支援本機及遠端連接埠轉送）"
    }
  ],
  "paths": {
...
    "/api/medicine/recognize": {
      "post": {
        "tags": [
          "Medicine"
        ],
        "summary": "Recognize a medicine image with the local model",
        "description": "Uploads one JPEG, PNG, or WebP image (maximum 10 MiB). The configured local recognizer returns up to three medicine license_number values; each is looked up in MariaDB using the same data source as GET /api/medicine.",
        "operationId": "recognizeMedicineImage",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG, PNG, or WebP image; maximum 10 MiB."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recognition completed. medicine is null when an ID is not present in MariaDB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicineRecognitionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or unsupported image upload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Image exceeds 10 MiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Recognizer failed or returned invalid output.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Recognizer or medicine database is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Recognizer timed out.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "MedicineDatabaseRecord": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "license_number",
          "chinese_name",
          "english_name",
          "shape",
          "color",
          "score_line",
          "size",
          "imprint_1",
          "imprint_2",
          "image_url"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "藥品編號，對應資料表 medicines.id。",
            "example": 2
          },
          "license_number": {
            "type": "string",
            "description": "許可證字號。",
            "example": "內衛成製字第000386號"
          },
          "chinese_name": {
            "type": "string",
            "nullable": true,
            "description": "中文藥名。",
            "example": "建功丸"
          },
          "english_name": {
            "type": "string",
            "nullable": true,
            "description": "英文藥名。",
            "example": "CHENG KONG PILL"
          },
          "shape": {
            "type": "string",
            "nullable": true,
            "description": "形狀。",
            "example": "其他"
          },
          "color": {
            "type": "string",
            "nullable": true,
            "description": "顏色。",
            "example": "棕"
          },
          "score_line": {
            "type": "string",
            "nullable": true,
            "description": "切線／刻痕。",
            "example": "無"
          },
          "size": {
            "type": "string",
            "nullable": true,
            "description": "外觀尺寸，保留來源文字；來源未註明單位。",
            "example": "8"
          },
          "imprint_1": {
            "type": "string",
            "nullable": true,
            "description": "標註一。",
            "example": ""
          },
          "imprint_2": {
            "type": "string",
            "nullable": true,
            "description": "標註二。",
            "example": ""
          },
          "image_url": {
            "type": "string",
            "nullable": true,
            "description": "圖片連結；多個連結保留來源的 ;;; 分隔符號。",
            "example": "https://mcp.fda.gov.tw/insert/shapeImg/89db57ae-5c85-47b8-9d74-351ecad719e6?c=o"
          }
        },
        "description": "MariaDB medicines 資料表中的藥品原始資訊；欄位名稱與資料庫一致。"
      },
      "MedicineRecognitionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ok",
          "recognition",
          "medicines"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "recognition": {
            "type": "object",
            "required": [
              "ids"
            ],
            "properties": {
              "ids": {
                "type": "array",
                "maxItems": 3,
                "items": {
                  "type": "string"
                }
              },
              "status": {
                "type": "string"
              },
              "detected_features": {
                "type": "object"
              },
              "detection_source": {
                "type": "string"
              },
              "predictions": {
                "type": "array",
                "maxItems": 3,
                "items": {
                  "type": "object"
                }
              }
            }
          },
          "medicines": {
            "type": "array",
            "maxItems": 3,
            "items": {
              "type": "object",
              "required": [
                "license_number",
                "medicine"
              ],
              "properties": {
                "license_number": {
                  "type": "string"
                },
                "medicine": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/MedicineDatabaseRecord"
                    }
                  ],
                  "nullable": true
                }
              }
            }
          }
        }
      },
...
    }
  }
}